Api calls are now their own module
